test: split up test files
